Progress towards v0.03, including data source chooser, but this commit also includes compile error, still needs more work

+++ craft3data/src/com/hiveworkshop/wc3/mdl/render3d/EmitterView.java
@@ -1,7 +1,7 @@
 package com.hiveworkshop.wc3.mdl.render3d;
 
 public abstract class EmitterView {
-    public int currentEmission;
+    public float currentEmission;
 
     public RenderModel instance;
